+++ src/POData/BatchProcessor/QueryParser.php
@@ -3,10 +3,8 @@
 declare(strict_types=1);
 namespace POData\BatchProcessor;
 
-use Illuminate\Http\Request;
 use POData\BaseService;
 use POData\OperationContext\ServiceHost;
-use POData\OperationContext\Web\Illuminate\IlluminateOperationContext;
 
 /**
  * Class QueryParser.